+++ cbmonitor/src/dashboards/single/analytics.ts
