+++ cbmonitor/src/dashboards/single/clusterManager.ts
