+++ cbmonitor/src/dashboards/single/fts.ts
