+++ cbmonitor/src/dashboards/single/index.ts
