+++ cbmonitor/src/dashboards/single/kv.ts
