+++ cbmonitor/src/dashboards/single/query.ts
