+++ cbmonitor/src/dashboards/single/system.ts
