+++ cbmonitor/src/dashboards/single/xdcr.ts
